+++ a38/__init__.py
@@ -1,0 +1,1 @@
+from .fattura import *  # noqa